% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{choropleth}
\alias{choropleth}
\title{Plot choropleth map of life expectancy}
\usage{
choropleth(data, quartile, sex, year, ra = TRUE)
}
\arguments{
\item{data}{A data frame containing life expectancies by state, year, and wealth quartile}

\item{quartile}{The income quartile to filter on (1, 2, 3, or 4)}

\item{sex}{A sex to filter on ("m" for male, "f" for female, or "mixed" for both)}

\item{year}{A particular year to filter on}

\item{ra}{(optional) A race adjusted filter (TRUE for race-adjusted, FALSE for race-unadjusted)}
}
\value{
A choropleth map of life expectancy across states for a given year and sex
}
\description{
Plot choropleth map of life expectancy
}
\details{
If provided, the year or years must fall within the range of years for which
there is data within the selected data frame.
}
\examples{
choropleth(health_ineq, quartile = 4, sex = "mixed", year = 2005)
choropleth(health_ineq, quartile = 1, sex = "m", year = 2001, ra = FALSE)
}
