% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wage_le_plot.R
\name{wage_le_plot}
\alias{wage_le_plot}
\title{Plot life expectancy by minimum wage}
\usage{
wage_le_plot(data, quartile, state = NULL, year = NULL, sex = NULL)
}
\arguments{
\item{data}{A data frame containing life expectancies by state, year, wealth quartile, and minimum wage.}

\item{quartile}{The income quartile to filter on (1, 2, 3, or 4).}

\item{state}{(optional) A state or set of states to filter on (abbreviated state names; default: all states).}

\item{year}{(optional) A year or set of years to filter on (default: all years).}

\item{sex}{(optional) A sex to filter on ("m", "f", or "mixed"; default: mixed).}
}
\value{
A scatterplot of life expectancy by minimum wage.
}
\description{
Plot life expectancy by minimum wage
}
\details{
If provided, the year or years must fall within the range of years for which
there is data within the selected data frame.
}
\examples{
wage_le_plot(health_ineq, quartile = 2, state = "MA", year = c(2000, 2001))
wage_le_plot(health_ineq, quartile = 4, year = 2013, state = "CT", sex = "f")
}
